Fix bug where edit patients changed predicate state
