+++ src/main/java/seedu/address/logic/commands/ListCommand.java
@@ -6,6 +6,7 @@ import static seedu.address.model.Model.PREDICATE_SHOW_ALL_PERSONS;
 
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
+import seedu.address.model.person.HiddenPersonPredicateSingleton;
 
 /**
  * Lists all persons in the address book to the user.
@@ -32,7 +33,7 @@ public class ListCommand extends Command {
         if (this.type.equals("patients")) {
             requireNonNull(model);
             model.updateFilteredPersonList(PREDICATE_SHOW_ALL_PERSONS);
-            ModelManager.hiddenPersons.clear();
+            HiddenPersonPredicateSingleton.clearHiddenList();
             return new CommandResult(MESSAGE_SUCCESS_PATIENTS);
         } else {
             requireNonNull(model);
