+++ src/main/java/seedu/address/model/person/TagContainsKeywordsPredicate.java
@@ -5,7 +5,6 @@ import java.util.List;
 import java.util.Set;
 import java.util.function.Predicate;
 
-import seedu.address.model.ModelManager;
 import seedu.address.model.tag.Tag;
 
 /**
@@ -24,7 +23,7 @@ public class TagContainsKeywordsPredicate implements Predicate<Person> {
         Set<Tag> tags = person.getTags();
         for (Tag t: tags) {
             if (tagKeywords.stream().anyMatch(t.tagName::equalsIgnoreCase)) {
-                ModelManager.hiddenPersons.add(person);
+                HiddenPersonPredicateSingleton.addToHiddenList(person);
                 return true;
             }
         }
